verify password when delete

+++ src/main/java/com/ryo/dgswcinside/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.ryo.dgswcinside.service;
 
 import com.ryo.dgswcinside.common.BaseResponse;
+import com.ryo.dgswcinside.dto.request.DeleteRequest;
 import com.ryo.dgswcinside.dto.request.PublishCommentRequest;
 import com.ryo.dgswcinside.dto.request.UpdateCommentRequest;
 import com.ryo.dgswcinside.entity.Comment;
@@ -20,25 +21,21 @@ public class CommentService {
         Post post = postRepository.findById(req.postId())
                 .orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다"));
         
-        Comment comment = new Comment();
-        comment.setPost(post);
-        comment.setPassword(req.password());
-        comment.setNickname(req.nickname());
-        comment.setContent(req.content());
-        
+        Comment comment = Comment.of(req, post);
         repository.save(comment);
         return BaseResponse.ok("댓글 등록 성공");
     }
-    
-    public BaseResponse<String> delete(Long id) {
+
+    public BaseResponse<String> delete(Long id, DeleteRequest req) {
+        if (!req.password().equals(repository.findById(id).orElseThrow().getPassword())) throw new RuntimeException("잘못된 비밀번호");
         repository.deleteById(id);
         return BaseResponse.ok("댓글 삭제 성공");
     }
-    
+
     public BaseResponse<String> update(Long id, UpdateCommentRequest req) {
         Comment comment = repository.findById(id).orElseThrow();
         if (!comment.getPassword().equals(req.password())) throw new RuntimeException("잘못된 비밀번호");
-        comment.setContent(req.content());
+        comment.update(req);
         repository.save(comment);
         return BaseResponse.ok("댓글 수정 성공");
     }
